agentic_research_assistant/requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.14
langchain-community>=0.2.0
crewai>=0.28.8
autogen-agentchat>=0.2.33
pandas>=2.2.0
scikit-learn>=1.4.0
spacy>=3.7.0
duckduckgo-search>=5.3.0
reportlab>=4.0.5
pypandoc>=1.13
python-dotenv>=1.0.1
copyleaks>=4.0.0
PyMuPDF>=1.24.0
requests>=2.32.0
beautifulsoup4>=4.12.0
markdown>=3.6
pydantic>=2.7.0
tiktoken>=0.7.0
